실시간 시정 출력 그래프 소스코드 모듈화
